Remove the "time" type, it is not useful.
Seriously, we added this because I thought that it
was actually a "duration". Like recording how long 
it took to cook your rice or somesuch.  It would 
be input in Hours::Minutes::Seconds (with maybe 
more precision options available).  However, 
that's not what this is, it is a timestamp with a 
required date attached to it, in which case you 
might as well use a DateTime which makes more 
sense anyway.
Duration would be seriously awesome, since it
wouldn't require the dev to do in-software 
conversions from the UI into seconds (or whatever
your lowest common denominator of precision is). 
Then the DB could do group queries on that stuff.
It would be cool. :)